refactor: centralize project saving
