+++ src/types.ts
@@ -1,0 +1,8 @@
+export interface PatternDetails {
+  grid: string[][];
+  fabricCount: number;
+  widthIn: number;
+  heightIn: number;
+  colors: string[];
+  colorUsage: Record<string, number>;
+}
